Fix/double to decimal (#2)
* undisabled checkStatus

* retyped double to big decimal

+++ src/main/java/com/cpallas/expenses/UserSession.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -12,6 +14,6 @@ public class UserSession {
 
     private Step step;
     private String description;
-    private Double amount;
+    private BigDecimal amount;
     private CategoryId categoryId;
 }
